Chemistry Calculator C-C++
The code involves the use of structs to assign data to different elements of the periodic table and also vastly involves the use of conditional statements in many of its constituent parts. It outputs a selection of 9 options available in the program. It then prompts the user to input a number in order to choose one of the said options. Once the user selects an option, a set of new instructions will appear and will ask the user to input known values/ data that can be used to find an answer using the inputted data. In some cases, the user may need to select an even more specific option in the calculator before inputting values as there is such a wide variety of options within the program. Once the user has inputted all the necessary information, the program will output a calculated answer and the program will finish running. If the user were to pick another option in the calculator, they will have to unfortunately run the program again and select the desired option.