Add files via upload
Chemistry is a subject that is very useful and applicable in everyday life. However, chemistry can be intimidating at times. Knowing that so many students are struggling to solve chemistry problems in a short amount of time, my colleagues and I created a chemistry calculator to assist all chemistry students in solving chemistry calculations more quickly and efficiently. 

Creating the calculator in C++ by implementing the use of structs and conditional statements. Implementing some of the chemistry formulas into the program so that users can select the calculation they want.